libtvg: Add built-in support for nodes.
Instead of implementing node / label support in pytvg.py, move this feature
to the builtin libtvg core. The idea is that each node can have an arbitrary
number of attributes. They can either be loaded from a file, or from the
same MongoDB that is used to load the graph.

Other changed:
* Use 64-bit integers for refcount.
* Ensure that there is at most one Python obj for each libtvg obj.